Moved GPT response formatting functions into GPTRequest.java, cleaned up GPTCrossOver.java
